% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedDataFrame-class.R,
%   R/DelayedDataFrame-method.R
\docType{class}
\name{DelayedDataFrame}
\alias{DelayedDataFrame}
\alias{.DelayedDataFrame}
\alias{DelayedDataFrame-class}
\alias{DelayedDataFrame}
\alias{coerce}
\alias{coerce,DataFrame,DelayedDataFrame-method}
\alias{lazyIndex<-}
\alias{[,DelayedDataFrame,ANY,ANY-method}
\alias{[,DelayedDataFrame-method}
\title{DelayedDataFrame-class}
\usage{
DelayedDataFrame(..., row.names = NULL, check.names = TRUE)

lazyIndex(x) <- value

\S4method{[}{DelayedDataFrame,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{...}{other arguments to pass.}

\item{row.names}{rownames}

\item{check.names}{if check names.}

\item{x}{input}

\item{i}{row subscript}

\item{j}{col subscript}

\item{drop}{if drop with reduced dimension, default is TRUE.}
}
\description{
The \code{DelayedDataFrame} class extends the
    \code{DataFrame} class and supports the storage of any type of
    object (with ‘length’ and ‘[’ methods) as columns.
}
\details{
The \code{DelayedDataFrame} inherits from \code{DataFrame}
    and behaves very similarily in terms of construction,
    subsetting, splitting, combining, etc. The most notable
    exception is that The additional slot of \code{lazyIndex},
    enables \code{DelayedArray} (with different back-ends) columns
    to share indexes when possible.
}
