% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedDataFrame-method.R
\docType{methods}
\name{as.list,DelayedDataFrame-method}
\alias{as.list,DelayedDataFrame-method}
\alias{names,DelayedDataFrame-method}
\alias{cbind,DelayedDataFrame-method}
\alias{bindROWS,DelayedDataFrame-method}
\alias{extractROWS,DelayedDataFrame,ANY-method}
\alias{extractROWS,DelayedDataFrame-method}
\title{\code{cbind} for DelayedDataFrame inherits the lazyIndex's if
inputs have any DelayedDataFrame objects. Otherwise, return a new
DelayedDataFrame with NULL lazyIndexes.}
\usage{
\S4method{as.list}{DelayedDataFrame}(x, use.names = TRUE)

\S4method{names}{DelayedDataFrame}(x)

\S4method{cbind}{DelayedDataFrame}(..., deparse.level = 1)

\S4method{bindROWS}{DelayedDataFrame}(x, objects = list(), use.names = TRUE,
  ignore.mcols = FALSE, check = TRUE)

\S4method{extractROWS}{DelayedDataFrame,ANY}(x, i)
}
\arguments{
\item{x}{a \code{DelayedDataFrame} object.}

\item{use.names}{whether to use the colnames of
\code{DelayedDataFrame} as the names for the returned list.}

\item{...}{One or more vector-like or matrix-like objects. These
can be given as named arguments.}

\item{deparse.level}{See ‘?base::cbind’ for a description of this
argument.}
}
\value{
colnames of \code{DelayedDataFrame}
}
\description{
DelayedDataFrame related methods. \code{as.list},
    \code{rbind} would incur realization.

\code{cbind} for DelayedDataFrame inherits the lazyIndex's if
inputs have any DelayedDataFrame objects. Otherwise, return a new
DelayedDataFrame with NULL lazyIndexes.

bindROWS is the lower-level function for \code{rbind}.
}
