% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyIndex-class.R
\docType{methods}
\name{bindROWS}
\alias{bindROWS}
\alias{bindCOLS,LazyIndex-class}
\title{bindROWS}
\usage{
\S4method{bindROWS}{LazyIndex}(x, objects = list(), use.names = TRUE,
  ignore.mcols = FALSE, check = TRUE)
}
\description{
bindROWS
}
