% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyIndex-class.R
\name{cbind}
\alias{cbind}
\alias{.cbind_lazyIndex}
\alias{bindCOLS,LazyIndex-class}
\title{cbind}
\usage{
.cbind_lazyIndex(x, objects = list(), use.names = TRUE,
  ignore.mcols = FALSE, check = TRUE)
}
\description{
cbind
}
